requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
astroid==3.0.2
attrs==23.2.0
backports.zoneinfo==0.2.1
bleach==6.1.0
cffi==1.16.0
cfgv==3.4.0
coverage==7.4.0
cryptography==42.0.1
dill==0.3.7
distlib==0.3.8
dj-database-url==2.1.0
Django==4.2.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-api-logger==1.1.15
drf-spectacular==0.27.1
exceptiongroup==1.2.0
filelock==3.13.1
flake8==7.0.0
gunicorn==21.2.0
identify==2.5.33
importlib-resources==6.1.1
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
isort==5.13.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mccabe==0.7.0
nodeenv==1.8.0
packaging==23.2
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
pluggy==1.4.0
pre-commit==3.5.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
pylint==3.0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pytest==7.4.4
pytest-cov==4.1.0
pytest-django==4.7.0
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
rpds-py==0.17.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0
uritemplate==4.1.1
virtualenv==20.25.0
webencodings==0.5.1
zipp==3.17.0
